statitics/requirements.txt
matplotlib==3.2.1
plotly==4.8.1
plotly_express==0.4.1
chart-studio==1.1.0
pillow==7.2.0
cufflinks==0.17.3
